services/poi-service/requirements.txt
fastapi
uvicorn
pydantic
openai
httpx
psycopg2-binary
asyncpg
python-dotenv
